+++ src/main/java/org/summerboot/jexpress/boot/annotation/HealthCheck.java
@@ -30,7 +30,7 @@ import java.lang.annotation.Target;
 @Retention(value = RetentionPolicy.RUNTIME)
 @Documented
 @BindingAnnotation
-public @interface Inspector {
+public @interface HealthCheck {
 
-    String name() default "";
+    String name();
 }