+++ src/main/java/org/summerboot/jexpress/security/SecurityUtil.java
@@ -151,6 +151,13 @@ public class SecurityUtil {
             return false;
         }
 
+        if (!sanitizePath(filePath) || !filePath.equals(realPath)) {
+            //var e = new ServiceError(appErrorCode, null, "⚠", null);
+            Err e = new Err(BootErrorCode.FILE_NOT_ACCESSABLE, null, "Invalid file path", null, "Malicious file request: " + filePath);
+            context.status(HttpResponseStatus.FORBIDDEN).error(e);
+            return false;
+        }
+
         if (!file.exists()) {
             //var e = new ServiceError(appErrorCode, null, "⚠", null);
             Err e = new Err(BootErrorCode.FILE_NOT_FOUND, null, "Invalid file path", null, "File not exists: " + filePath);
@@ -158,11 +165,10 @@ public class SecurityUtil {
             return false;
         }
 
-        if (!sanitizePath(filePath) || !filePath.equals(realPath)
-                || file.isDirectory() || !file.isFile()
+        if (file.isDirectory() || !file.isFile()
                 || file.isHidden() || !file.canRead()) {
             //var e = new ServiceError(appErrorCode, null, "⚠", null);
-            Err e = new Err(BootErrorCode.FILE_NOT_ACCESSABLE, null, "Invalid file path", null, "Malicious file request: " + filePath);
+            Err e = new Err(BootErrorCode.FILE_NOT_ACCESSABLE, null, "Invalid file path", null, "Can not read file: " + filePath);
             context.status(HttpResponseStatus.FORBIDDEN).error(e);
             return false;
         }